Added parts of the back-end to the userlist
- Added part of the backend to hte userlist, had to create a new ajax function to call in the userlist and the delete user process.

+++ includes/views/userlist.php
@@ -19,34 +19,67 @@ if (isset($_GET['s'])) {
 	$search  = $_GET['s'];
 	$category = $_GET['c'];
 
-    $users = User::find_user($search, $category);
+	$users = User::find_user($search, $category);
 
 } else {
 
-  $search = "";
-  $users = User::find_all();
+	$search = "";
+	$users = User::find_all();
 
 }
 
 ?>
 
-	<table>
-		<thead>
 
-			<tr>
-				<th>Image</th>
-				<th>Username</th>
-				<th>First Name</th>
-				<th>Last Name</th>
-				<th>Joined</th>
-			</tr>
 
-		</thead>
-		<tbody>
+<table class="user-list-list">
+	<thead>
+		<tr>
+			<th>Avatar</th>
+			<th>Username</th>
+			<th class="user-list-hide-laptop">Name</th>
+			<th class="user-list-hide-tablet">Email</th>
+			<th>Actions</th>
+		</tr>
+	</thead>
+
+	<tbody>
+
 
 		<!-- For each loop that runs trough all the elements in the $games array. -->
-		<?php foreach ($users as $user) : ?>
+		<?php foreach ($users as $user) : 
 
+			// Gets the username and then makes the letters lowercase.
+			$username = strtolower($user->username);
+			$first_name = strtolower($user->first_name);
+			$middle_name = strtolower($user->middle_name);
+			$last_name = strtolower($user->last_name);
+			$full_name = $first_name . " " . $middle_name . " " . $last_name;
+
+		?>
+
+
+
+		<tr>
+			<td><div class="avatar -s" style="background-image: url(<?php echo $user->get_user_image(); ?>)"></div></td>
+			<td><?php echo str_replace($search, '<mark>' . $search . '</mark>', $username); ?></td>
+			<td class="user-list-hide-laptop"><?php echo $full_name; ?></td>
+			<td class="user-list-hide-tablet"><?php echo str_replace($search, '<mark>' . $search . '</mark>', $user->email); ?></td>
+			<td>
+				<div class="user-list-actions">
+					<a href="profile.php?id=<?php echo $user->id ?>" data-tooltip="Profile" class="user-list-action tooltip"><i class="fas fa-user"></i></a>
+
+					<!-- Legg till en sjekk for om de virkelig vil slette denne brukeren -->
+					<button data-tooltip="Delete" class="user-list-action -delete tooltip" 
+						onclick="return confirm('Are you sure?')?delete_user(<?php echo $user->id; ?>):'';"><i class="fas fa-user-times"></i></button>
+
+					<!-- Legg till en sjekk for om de virkelig vil promote denne brukeren -->
+					<button data-tooltip="Promote" class="user-list-action -promote tooltip"><i class="fas fa-user-tie"></i></button>
+				</div>
+			</td>
+		</tr>
+
+<!--
 			<tr>
 
 				<td>
@@ -58,7 +91,7 @@ if (isset($_GET['s'])) {
 				<td>
 					<a href="profile.php?id=<?php echo $user->id; ?>">
 						
-					<?php 
+						<?php 
 
 						// Gets the username and then makes the letters lowercase.
 						$username = strtolower($user->username);
@@ -66,17 +99,17 @@ if (isset($_GET['s'])) {
 						// Switches out what is searched with the same with marks around it.
 						echo str_replace($search, '<mark>' . $search . '</mark>', $username); 
 
-					?>
-					
+						?>
+
 					</a>
 				</td>
 
 				<td>
 					<?php 
 						// Gets the username and then makes the letters lowercase.
-						$first_name = strtolower($user->first_name);
+					$first_name = strtolower($user->first_name);
 						// Switches out what is searched with the same with marks around it.
-						echo str_replace($search, '<mark>' . $search . '</mark>', $first_name); 
+					echo str_replace($search, '<mark>' . $search . '</mark>', $first_name); 
 					?>
 				</td>
 
@@ -90,12 +123,14 @@ if (isset($_GET['s'])) {
 				</td>
 
 				<td><?php echo $user->joined; ?></td>
-				<td><a href="includes/process/delete_user.php?id=<?php echo $user->id; ?>" 
-					onclick="return confirm('Are you sure you want to delete <?php echo $user->username ?>?')">Delete</a></td>
-				<td><a href="profile.php?id=<?php echo $user->id; ?>">View</a></td>
+				<td><a href="includes/process/delete_user.php?id=<?php echo $user->id; ?>" onclick="return confirm('Are you sure you want to delete <?php echo $user->username ?>?')">Delete</a></td>
+					<td><a href="profile.php?id=<?php echo $user->id; ?>">View</a></td>
 
-			</tr>
+				</tr>
 
-		<?php endforeach; ?>
+-->
+
+			<?php endforeach; ?>
+
 		</tbody>
 	</table>
